Move non-distributable layer functions to own file
